+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+# v1.1.0
+
+* Added keybind for toggling the rainbow (`F9` by default)
+* Added config file for changing toggle key and setting cycle speed
+
 # v1.0.0
-Initial release
+
+* Initial release